refactor: update module path and improve worker pool implementation
- Change module path from 'go-image-inspector' to 'github.com/anime-shed/image-inspector-go'
- Enhance worker pool with better resource management and error handling
- Add SSRF protection and security improvements to HTTP image fetcher
- Optimize metrics calculator with better parallel processing and edge case handling

+++ internal/analyzer/worker_pool_test.go
@@ -169,11 +169,11 @@ func TestWorkerPool_AtomicCounters(t *testing.T) {
 	stats := pool.GetStats()
 
 	// Verify counters
-	if stats.TotalJobs != numJobs {
+	if stats.TotalJobs != int64(numJobs) {
 		t.Errorf("Expected %d total jobs, got %d", numJobs, stats.TotalJobs)
 	}
 
-	if stats.CompletedJobs != numJobs {
+	if stats.CompletedJobs != int64(numJobs) {
 		t.Errorf("Expected %d completed jobs, got %d", numJobs, stats.CompletedJobs)
 	}
